backend/decision-service/requirements.txt
paho-mqtt==2.1.0
scikit-fuzzy==0.5.0
numpy==2.2.4
